wrapper/src/services/python-vnc/requirements.txt
mss
pillow
pynput
httpx
websockets
